session 11

+++ app/Http/Controllers/AuthorController.php
@@ -12,6 +12,9 @@ class AuthorController extends Controller
     //
 
     public function store(Request $request){
+        if(!$request->user()->can("create authors") /* return true or false */){
+            return $this->responseForbidden("create authors", "You need to have create authors permission to proceed!");
+        }
         $validator = Validator::make($request->all(), [
             'firstname' => "sometimes|max:64|regex:/^[a-z ,.'-]+$/i",
             'middlename' => "sometimes|max:64|regex:/^[a-z ,.'-]+$/i",
@@ -24,6 +27,7 @@ class AuthorController extends Controller
         }
 
         $author = Author::create($validator->validated());
+        $this->log($request, "create author", $validator->validated(), "authors", $author);
         Cache::forget("authors");
         return $this->responseCreated($author, "Author has been created!");
         
@@ -71,6 +75,9 @@ class AuthorController extends Controller
     }
 
     public function update(Request $request, Author $author){
+        if(!$request->user()->can("update authors") /* return true or false */){
+            return $this->responseForbidden("update authors", "You need to have update authors permission to proceed!");
+        }
         $validator = Validator($request->all(), [
             'firstname' => "sometimes|max:64|regex:/^[a-z ,.'-]+$/i",
             'middlename' => "sometimes|max:64|regex:/^[a-z ,.'-]+$/i",
@@ -80,13 +87,31 @@ class AuthorController extends Controller
         if($validator->fails()){
             return $this->responseBadRequest($validator);
         }
-
+        $orig = $author->toArray();
         $author->update($validator->validated());
+        $dirtyorig = array(); // empty array of list that will be changed
+        $dirtynew = array();
+        // dd($validator->validated(), $orig);
+        foreach($validator->validated() as $key => $value){
+            if($value != $orig[$key]){
+                $dirtyorig[$key] = $orig[$key]; // columns that has been changed
+                $dirtynew[$key] = $value; // columns that has been changed
+            }
+        }
+        $this->log($request, "update author", [
+            "old" => $dirtyorig,
+            "new" => $dirtynew
+        ], "authors", $author);
         Cache::forget("authors");
         return $this->responseOk($author, "Author has been updated!");
     }
 
-    public function destroy(Author $author){
+    public function destroy(Request $request, Author $author){
+        if(!$request->user()->can("delete authors") /* return true or false */){
+            return $this->responseForbidden("delete authors", "You need to have delete authors permission to proceed!");
+        }
+        $orig = $author->toArray();
+        $this->log($request, "delete author", $orig, "authors", $author);
         $author->books()->delete();
         $author->delete();
         Cache::forget("authors");
